requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

Flask-Migrate==4.0.5

python-jose==3.3.0
requests==2.31.0

gunicorn==21.2.0

auth0-server-python>=1.0.0b7
flask[async]>=2.0.0
